Entrega/Task 2/requirements.txt
pandas>=1.3.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
PyYAML>=5.4.0
